+++ src/directives/picto.js
@@ -10,7 +10,9 @@
          * @scope
          * @restrict E
          * @param {string} url The url of the svg or image to display
-         * @param {string} color The color to use to fill the svg
+         * @param {string} color The color to use to fill the SVG
+         * @param {Object} colorByName An object containing a mapping between element names within the SVG, and colors.
+         * The elements within the SVG with a matching `name` attribute will take the corresponding color.
          * @param {string} classes The classes to directly apply to the svg element
          * @description
          * This widget displays a "picto" specified by a url and force a fill color on it.
@@ -33,13 +35,14 @@
             scope: {
                 url: '=',
                 color: '=',
+                colorByName: '=',
                 classes: '='
             },
 
             template: '<div class="odswidget odswidget-picto {{ classes }}"></div>',
             link: function(scope, element) {
                 var svgContainer;
-                scope.$watch('[url, color]', function(nv) {
+                scope.$watch('[url, color, colorByName]', function(nv) {
                     if (nv[0]) {
                         if (Modernizr && !Modernizr.svg) {
                             return;
@@ -47,7 +50,7 @@
                         if (svgContainer) {
                             element.empty();
                         }
-                        svgContainer = SVGInliner.getElement(scope.url, scope.color);
+                        svgContainer = SVGInliner.getElement(scope.url, scope.color, scope.colorByName);
                         if (!scope.color) {
                             svgContainer.addClass('ods-svginliner__svg-container--colorless');
                         }
@@ -80,7 +83,6 @@
             link: function(scope, element) {
                 scope.originalClasses = element.attr('class').replace('ng-isolate-scope', '').trim();
                 var template = '<ods-picto url="themeConfig.url" aria-label="Theme of this dataset: {{ theme|firstValue }}" translate="aria-label" color="themeConfig.color" classes="originalClasses + \' odswidget-theme-picto theme-\' + (getTheme()|themeSlug) "></ods-picto>';
-                var themeConfig = null;
                 var defaultPicto = false;
                 if (ODSWidgetsConfig.themes[scope.theme] && ODSWidgetsConfig.themes[scope.theme].url) {
                     scope.themeConfig = ODSWidgetsConfig.themes[scope.theme];