+++ src/services/api.js
@@ -151,7 +151,7 @@
                 // FIXME: Why don't we implicitely use the parameters from the context, instead of requiring the widgets
                 // to explicitely send them together with the other parameters?
                 'analyze': function(context, parameters, timeout) {
-//                    return request(context, '/api/datasets/1.0/'+context.dataset.datasetid+'/records/analyze/', parameters);
+                    // return request(context, '/api/datasets/1.0/'+context.dataset.datasetid+'/records/analyze/', parameters);
                     return request(context, '/api/records/1.0/analyze/', angular.extend({}, parameters, {dataset: context.dataset.datasetid}), timeout)
                         .success(function(data, status, headers, config) {
                             if (headers()['ods-analyze-truncated']) {